de-coupled the tower functionality
